Bashrc: set unlimited history size, no duplicates
Until now, my history file was limited and contained duplicates.

With this commit, the history will not store duplicate commands
anymore. However it will not remove the previous duplicates. For that, you
might need to remove the duplicates with a text-manipulation program.

Also, add the '-x' switch to the 'clear' command with an alias so that it
will not remove the entire terminal history.